# download/one-command-hospital/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Mediator API — Guideline Copilot
  version: 0.6.1
  description: |
    The ONE contract every future AI capability rides on: clinical question in,
    cited-and-verified (or refused) answer out, tamper-evident audit + FHIR
    writeback with provenance, human sign-off loop.

    Safety contracts embedded in this API (breaking them is a Class B change
    under docs/governance.md):
      - Questions are de-identified server-side before any model path.
      - Answers carry edition-stamped citations or the exact refusal marker
        NOT_COVERED.
      - A missing citation or an unreachable verifier marks grounding=false.
      - Every answer/audit event carries an X-Request-ID trace id.
servers:
  - url: http://localhost:8103
    description: sandbox compose (plaintext, inside lab)
  - url: https://localhost:8443/api
    description: sandbox edge TLS terminator (make edge-up)
paths:
...
  /process:
    post:
      summary: Ask the Guideline Copilot
      operationId: processQuestion
      parameters:
        - in: header
          name: X-Request-ID
          required: false
          description: Caller trace id; minted server-side when absent.
          schema: { type: string, maxLength: 64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question:
                  type: string
                  description: Clinical question. May contain narrative; it is
                    de-identified server-side before reaching any model.
                  maxLength: 100000
                contextText:
                  type: string
                  description: Optional note context; de-identified like the question.
                patientId:
                  type: [string, "null"]
                  description: FHIR Patient id for Communication.subject (id only).
                userId:
                  type: string
                  default: unknown
                  description: Requesting clinician id — audit actor.
      responses:
        "200":
          description: |
            Answer (or refusal) with full provenance. Component failure
            mid-pipeline marks grounding=false and human_action=pending_review
            — the answer is never presented as verified when it is not.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessResponse"
        "400": { description: Missing question }
        "429": { description: Rate limit exceeded }
        "502": { description: Pipeline failure before any answer could form }
  /signoff:
    post:
      summary: Clinician closes the human_action loop
      operationId: signoffCommunication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [communicationId, action]
              properties:
                communicationId: { type: string }
                action: { type: string, enum: [signed_off, escalated] }
                by: { type: string, default: clinician }
      responses:
        "200":
          description: Sign-off recorded (ok may be false if FHIR write failed)
        "400": { description: Missing/invalid fields }
        "404": { description: Communication not found in the FHIR spine }
        "502": { description: FHIR spine unreachable }
...
components:
  schemas:
...
    ProcessResponse:
      type: object
      properties:
        communication:
          type: object
          description: FHIR Communication resource (payload[0].contentString is
            the de-identified provenance envelope).
        fhir:
          type: object
          properties:
            written: { type: boolean }
            status: { type: integer }
            warning: { type: string }
        audit_seq:
          type: integer
          description: Sequence number of the audit record for this answer.
        total_latency_ms: { type: integer }